Major reorganisation.
The key point is that we define three custom GraphQL types : Article,
Seminar and Menu that are to be used in queries instead of the more
generic Mdx or File types.

In addition, each type is handled within its own theme and we make use
of
[shadowing](https://www.gatsbyjs.com/docs/how-to/plugins-and-themes/shadowing/#reach-skip-nav)
to more easily develop things in isolation. That should also ease the
transition to new styling in the future.

More on a detailed/practical level, note that :

- all component filenames are now consistently lowercase
- linting (using eslint and prettier) has been introduced as a mandatory
  step for commit